+++ spring-cloud-starter-tencent-polaris-circuitbreaker/src/main/java/com/tencent/cloud/polaris/circuitbreaker/reporter/CircuitBreakerPlugin.java
@@ -18,7 +18,6 @@
 package com.tencent.cloud.polaris.circuitbreaker.reporter;
 
 import com.tencent.cloud.common.constant.ContextConstant;
-import com.tencent.cloud.common.metadata.MetadataContext;
 import com.tencent.cloud.common.metadata.MetadataContextHolder;
 import com.tencent.cloud.polaris.circuitbreaker.PolarisCircuitBreaker;
 import com.tencent.cloud.polaris.circuitbreaker.instrument.resttemplate.PolarisCircuitBreakerHttpResponse;
@@ -28,7 +27,6 @@ import com.tencent.cloud.rpc.enhancement.plugin.EnhancedPluginType;
 import com.tencent.cloud.rpc.enhancement.plugin.EnhancedRequestContext;
 import com.tencent.cloud.rpc.enhancement.plugin.EnhancedResponseContext;
 import com.tencent.cloud.rpc.enhancement.plugin.reporter.SuccessPolarisReporter;
-import com.tencent.polaris.api.utils.StringUtils;
 import com.tencent.polaris.circuitbreak.client.exception.CallAbortedException;
 import com.tencent.polaris.metadata.core.MetadataType;
 import org.slf4j.Logger;
@@ -70,9 +68,9 @@ public class CircuitBreakerPlugin implements EnhancedPlugin {
 		EnhancedRequestContext request = context.getRequest();
 		EnhancedResponseContext response = context.getResponse();
 
-		String host = request.getServiceUrl() != null ? request.getServiceUrl().getHost() : request.getUrl().getHost();
-		String path = request.getServiceUrl() != null ? request.getServiceUrl().getPath() : request.getUrl().getPath();
-		String governanceNamespace = StringUtils.isNotEmpty(request.getGovernanceNamespace()) ? request.getGovernanceNamespace() : MetadataContext.LOCAL_NAMESPACE;
+		String host = request.getHost();
+		String path = request.getPath();
+		String governanceNamespace = request.getGovernanceNamespace();
 
 		String httpMethod = request.getHttpMethod().name();