+++ CHANGELOG.md
@@ -25,3 +25,4 @@
 - [feat:support polaris server nearby backup LB.](https://github.com/Tencent/spring-cloud-tencent/pull/1628)
 - fix:fix ipv6 instance not found bug.
 - fix:fix tsf lossless not working bug.
+- fix:fix RestTemplate trace attribute loss bug.